RCAN/requirements_.txt
cffi                     1.14.2
cmake                    3.18.0
contourpy                1.1.1
cycler                   0.10.0
decorator                4.4.2
filelock                 3.15.1
fonttools                4.53.0
fsspec                   2024.6.0
future                   0.18.2
imageio                  2.34.2
importlib                1.0.4
importlib_resources      6.4.0
intel-openmp             2020.0.133
Jinja2                   3.1.4
kiwisolver               1.1.0
lazy_loader              0.4
MarkupSafe               2.1.5
matplotlib               3.0.3
mkl                      2019.0
mkl-include              2019.0
mpmath                   1.3.0
networkx                 3.1
num3.py                    1.24.4
nvidia-cublas-cu12       12.1.3.1
nvidia-cuda-cupti-cu12   12.1.105
nvidia-cuda-nvrtc-cu12   12.1.105
nvidia-cuda-runtime-cu12 12.1.105
nvidia-cudnn-cu12        8.9.2.26
nvidia-cufft-cu12        11.0.2.54
nvidia-curand-cu12       10.3.2.106
nvidia-cusolver-cu12     11.4.5.107
nvidia-cusparse-cu12     12.1.0.106
nvidia-nccl-cu12         2.20.5
nvidia-nvjitlink-cu12    12.5.40
nvidia-nvtx-cu12         12.1.105
packaging                24.1
pandas                   2.0.3
pillow                   10.3.0
pip                      24.1
pycparser                2.20
pyparsing                2.4.7
python-dateutil          2.8.1
pytz                     2024.1
PyWavelets               1.1.1
PyYAML                   5.3.1
scikit-image             0.21.0
scipy                    1.10.1
setuptools               70.1.0
six                      1.15.0
sympy                    1.12.1
tifffile                 2023.7.10
torch                    2.3.1
torchvision              0.2.1
tqdm                     4.19.9
triton                   2.3.1
typing                   3.7.4.3
typing_extensions        4.12.2
tzdata                   2024.1
wheel                    0.43.0
zipp                     3.19.2
